Reorganized to monorepo, setup tsdx, stopped using yarn.
Created new `lib-core` (and placeholder `lib-gdrive`) directories that will be homes for separate packages and set them up with tsdx. Moved most of `src/*` to `lib-core/src/`. Fixed linting and jest errors in unit tests that became apparent with tsdx. Had to stop using yarn since it kept resolving to the older version of jest and ts-jest that came with tsdx instead of manually-installed jest 26.6+ (which is needed for calls like `jest.useFakeTimers('modern')`).