Modified plain JS example app to use IndexedDB API and UMD version of sidesync.
Modified `npm run build` to generate a UMD version of lib-core (and modified to use --name flag so that library has "sidesync" global var name) so that it can be used in the plain JS example app via <script src="..."> tag. Modified example app to use IndexedDB for storage, and to call sidesync.setupStores().