Added proxy implementations for store.delete(), get(), and getAll(), and put().
Calling .delete() on a proxied store will now do a "soft delete" and record the mutation in the oplog. Calling .getAll() on a proxied store will now filter out "soft deleted" items from the results. put() can now be called with only a subset of the full object properties (i.e., just the ones that should be updated).