Removes all support for automatic/invisible "soft deletion" of objects.
The idea for this feature was that upstream devs could call delete() on proxied object stores and cursors, and IDBSideSync would intercept those calls. It would then prevent objects from actually being deleted, instead _updating_ them with a special property that, if true, meant the object should be treated as if it were deleted.

It was eventually discovered that supporting this was a bad idea for a few reasons. First, it involved mutating the upstream dev's objects (i.e., adding a "deleted: true" property to their objects). More importantly, this implied trying to filter out "deleted" objects in every other case where the API was being used to read objects. As it turns out, it's not possible to support this "filter on read" approach in every situation.

For example, imagine an upstream dev calling `.count(keyRange)` on their object store, where `keyRange` was some IDBKeyRange instance for only including objects whose "age" property value was in some bounds. In that scenario, it's not possible to modify the keyRange to _include_ criteria for excluding objects with a "deleted" property by default.

For this to work, you would need to ensure that every index the upstream dev creates includes the "deleted" property (i.e., you'd have to intercept every call to create an index and modify it to be created with this key. This means you'd then have to intercept every call on that index to read something back out (e.g., count(key|keyRange), openCursor(key|keyRange), etc.), and then modify the key or keyRange arg to include the criteria for excluding items that have the "deleted" property set to some "truthy" value (i.e., "1" since you can't use booleans as keys).

It's theoretically possible to do this, but it would be a lot of work--and would still involve mutating the upstream dev's objects. For now, it seems better to have developers manage this themselves. It's not that much more work to update something with a "soft deleted" property, and arguably, it's something they should be forced to think about handling (especially if they want to support the ability to "undo" the deletion of an object).