Improved error checking in IDBObjectStoreProxy.proxiedPut() method, added comments.
Modified IDBObjectStore.proxiedPut() so that, given the "temp" put() and "permanent" put() calls that it makes to update objects in the IDBObjectStore, an `onsuccess` handler is assigned to BOTH requests. The "onsuccess" handler for the "temp" put sets a "temp put finished" flag; the "onsuccess" handler for the "permanent" put() checks the value of that flag when it completes, and throws an error if it finds that the temp put() hasn't completed yet. In this way, we are checking to ensure that the "permanent" put() runs last.