Improved Cypress tests.
Modified utils.ts:resolveOnTxComplete() so that it does not resolve until the callback it is passed also resolves (i.e., in addition to the db transaction completing). Previously, it was not waiting on the callback to resolve, which had unintended consequences for unit tests (e.g., a test could finish before the callback finished).