Renamed IDB object stores used by Cypress tests.
Made this change so that the store names actually described the things they store to make things a bit easier to read/understand. They also match the example app, which helps make the tests a bit easier to think about when comparing them to real-world scenarios that the example app demos.