Fixed bug with Cypress unit tests where cached IDBSideSync settings (from db.ts:initSettings()) were incorrectly being re-used between tests.
This refactors the db.ts:initSettings() function in a couple of ways.

First, it no longer returns the cachedSettings variable if it exists with a value. This was necessary because the tests cause initSettings() to be called multiple times without reloading the "db.ts" module between runs; in effect, the first test would set the cachedSettings variable, and all following tests would end up getting that same cached value--which would differ from what was actually stored in the database. This could cause tests that compare in-memory settings to database-persisted settings to fail.

Secondly, this commit refactors initSettings() to more explicitly/clearly use the promise API instead of async/await syntax sugar. It's much easier to actually see what's happening now without hiding the transaction handling code, and easier to see the _order_ in which it happens without using so many nested `await` statements.